backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.29
alembic==1.13.1
asyncpg==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
celery[redis]==5.3.6
redis==5.0.4
boto3==1.34.87
Pillow==10.3.0
opencv-python-headless==4.9.0.80
paddlepaddle
paddleocr
deskew==1.5.1
numpy==1.26.4
google-generativeai==0.5.4
openai==1.25.0
openpyxl==3.1.2
python-dotenv==1.0.1
sentry-sdk[fastapi]==1.45.0
aiofiles==23.2.1
PyMuPDF==1.24.1
pyafipws==0.0.2
authlib==1.3.0
itsdangerous==2.1.2
resend==0.7.2
